adjust: 添加自定义属性
