+++ RippleView/build.gradle
@@ -1,17 +1,20 @@
 plugins {
     id 'com.android.library'
     id 'kotlin-android'
+    id 'com.github.dcendents.android-maven'
 }
 
+group = 'com.github.Leo199206'
+version = "1.0.03"
 android {
     compileSdkVersion 30
-    buildToolsVersion "30.0.3"
+    buildToolsVersion "30.0.1"
 
     defaultConfig {
         minSdkVersion 21
         targetSdkVersion 30
         versionCode 1
-        versionName "1.0"
+        versionName version
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         consumerProguardFiles "consumer-rules.pro"
@@ -30,15 +33,23 @@ android {
     kotlinOptions {
         jvmTarget = '1.8'
     }
+    /**
+     * 配置编译资源目录
+     */
+    sourceSets {
+        debug.setRoot('src/debug')
+        release.setRoot('src/release')
+        main {
+            //配置java和kotlin源码目录
+            java.srcDirs = ['src/main/kotlin', 'src/main/java']
+            //指定jniLibs文件存放目录（.so文件）
+            jniLibs.srcDirs = ['libs']
+        }
+    }
 }
 
 dependencies {
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation 'androidx.core:core-ktx:1.3.2'
-    implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation 'com.google.android.material:material:1.2.0'
-    testImplementation 'junit:junit:4.+'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
 }
