--- RippleView/src/main/java/com/ripple/view/RippleCircle.kt
@@ -1,11 +1,0 @@
-package com.ripple.view
-
-/**
- * <pre>
- *   @author : leo
- *   @time   : 2021/03/21
- *   @desc   : 圆参数信息
- * </pre>
- */
-class RippleCircle(var radius: Float, var alpha: Int) {
-}
