+++ RippleView/src/main/kotlin/com/ripple/view/RippleCircle.kt
@@ -1,0 +1,11 @@
+package com.ripple.view
+
+/**
+ * <pre>
+ *   @author : leo
+ *   @time   : 2021/03/21
+ *   @desc   : 圆参数信息
+ * </pre>
+ */
+class RippleCircle(var radius: Float, var alpha: Int) {
+}
