+++ RippleView/src/main/res/values/attrs.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <declare-styleable name="RippleView">
+        <attr name="ripple_circle_color" format="color" />
+        <attr name="ripple_circle_min_radius" format="dimension" />
+        <attr name="ripple_circle_count" format="integer" />
+        <attr name="ripple_circle_style" format="enum">
+            <enum name="FILL" value="0" />
+            <enum name="STROKE" value="1" />
+        </attr>
+        <attr name="ripple_speed" format="float" />
+        <attr name="ripple_circle_stroke_width" format="dimension" />
+    </declare-styleable>
+</resources>
