+++ app/src/main/java/com/ripple/sample/MainActivity.kt
@@ -4,10 +4,14 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import kotlinx.android.synthetic.main.activity_main.*
 
+/**
+ * RippleView Sample
+ */
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-        ripple.onStart(this)
+        ripple_stroke.onStart(this)
+        ripple_fill.onStart(this)
     }
 }
